+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/bean/entity/HibernateImageListNode.java
@@ -1,5 +1,7 @@
 package com.dwarfeng.settingrepo.impl.bean.entity;
 
+import com.dwarfeng.datamark.bean.jpa.DatamarkEntityListener;
+import com.dwarfeng.datamark.bean.jpa.DatamarkField;
 import com.dwarfeng.settingrepo.sdk.util.Constraints;
 import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
@@ -12,9 +14,10 @@ import java.util.Set;
 @Entity
 @IdClass(HibernateStringIdKey.class)
 @Table(name = "tbl_image_list_node")
+@EntityListeners(DatamarkEntityListener.class)
 public class HibernateImageListNode implements Bean {
 
-    private static final long serialVersionUID = 944941176607547149L;
+    private static final long serialVersionUID = 1552020344143639460L;
 
     // -----------------------------------------------------------主键-----------------------------------------------------------
     @Id
@@ -24,6 +27,22 @@ public class HibernateImageListNode implements Bean {
     // -----------------------------------------------------------主属性字段-----------------------------------------------------------
     @Column(name = "size", nullable = false)
     private int size;
+
+    // -----------------------------------------------------------审计-----------------------------------------------------------
+    @DatamarkField(handlerName = "settingNodeDatamarkHandler")
+    @Column(
+            name = "created_datamark",
+            length = com.dwarfeng.datamark.util.Constraints.LENGTH_DATAMARK_VALUE,
+            updatable = false
+    )
+    private String createdDatamark;
+
+    @DatamarkField(handlerName = "settingNodeDatamarkHandler")
+    @Column(
+            name = "modified_datamark",
+            length = com.dwarfeng.datamark.util.Constraints.LENGTH_DATAMARK_VALUE
+    )
+    private String modifiedDatamark;
 
     // -----------------------------------------------------------一对多-----------------------------------------------------------
     @OneToMany(cascade = CascadeType.MERGE, targetEntity = HibernateImageListNodeItem.class, mappedBy = "node")
@@ -58,6 +77,22 @@ public class HibernateImageListNode implements Bean {
         this.size = size;
     }
 
+    public String getCreatedDatamark() {
+        return createdDatamark;
+    }
+
+    public void setCreatedDatamark(String createdDatamark) {
+        this.createdDatamark = createdDatamark;
+    }
+
+    public String getModifiedDatamark() {
+        return modifiedDatamark;
+    }
+
+    public void setModifiedDatamark(String modifiedDatamark) {
+        this.modifiedDatamark = modifiedDatamark;
+    }
+
     public Set<HibernateImageListNodeItem> getItems() {
         return items;
     }
@@ -70,6 +105,8 @@ public class HibernateImageListNode implements Bean {
     public String toString() {
         return getClass().getSimpleName() + "(" +
                 "stringId = " + stringId + ", " +
-                "size = " + size + ")";
+                "size = " + size + ", " +
+                "createdDatamark = " + createdDatamark + ", " +
+                "modifiedDatamark = " + modifiedDatamark + ")";
     }
 }