+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/configuration/ServiceExceptionMapperConfiguration.java
@@ -17,6 +17,7 @@ public class ServiceExceptionMapperConfiguration {
     public MapServiceExceptionMapper mapServiceExceptionMapper() {
         Map<Class<? extends Exception>, ServiceException.Code> destination = ServiceExceptionHelper.putDefaultDestination(null);
         destination = com.dwarfeng.ftp.util.ServiceExceptionHelper.putDefaultDestination(destination);
+        destination = com.dwarfeng.datamark.util.ServiceExceptionHelper.putDefaultDestination(destination);
         destination.put(FormatterException.class, ServiceExceptionCodes.FORMATTER_FAILED);
         destination.put(FormatterExecutionException.class, ServiceExceptionCodes.FORMATTER_EXECUTION_FAILED);
         destination.put(FormatterMakeException.class, ServiceExceptionCodes.FORMATTER_MAKE_FAILED);